server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Muinde"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.79.0"
uvicorn = "^0.18.2"
alembic = "^1.8.1"
psycopg2-binary = "^2.9.3"
tenacity = "^8.0.1"
python-dotenv = "^0.20.0"
autoflake = "^1.4"
black = "^22.6.0"
mypy = "^0.971"
SQLAlchemy-Utils = "^0.38.3"
SQLAlchemy-Searchable = "^1.4.1"
fastapi-pagination = "^0.9.3"
tabulate = "^0.8.10"
ipython = "^8.4.0"
isort = "^5.10.1"
passlib = "^1.7.4"
python-jose = "^3.3.0"
emails = "^0.6"
Jinja2 = "^3.1.2"
pydantic = {extras = ["email"], version = "^1.9.1"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
